feat: update homepage and add navigation wrapper
